Added start and end date parsing to Pingdom reports.

+++ src/main/java/com/onsdigital/performance/reporter/pingdom/model/SummaryResponseTime.java
@@ -1,7 +1,7 @@
 package com.onsdigital.performance.reporter.pingdom.model;
 
 public class SummaryResponseTime {
-    public int from;
-    public int to;
-    public int avgresponse;
+    public long from;
+    public long to;
+    public long avgresponse;
 }
