+++ src/main/java/com/onsdigital/performance/reporter/pingdom/model/SummaryStatus.java
@@ -1,7 +1,7 @@
 package com.onsdigital.performance.reporter.pingdom.model;
 
 public class SummaryStatus {
-    public int totalup;
-    public int totaldown;
-    public int totalunknown;
+    public long totalup;
+    public long totaldown;
+    public long totalunknown;
 }
